# docs/swagger.yaml
basePath: /api/v1
definitions:
  diprec_api_internal_domain.AuthResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/diprec_api_internal_domain.UserResponse'
    type: object
  diprec_api_internal_domain.CourseResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  diprec_api_internal_domain.Error:
    properties:
      message:
        type: string
    type: object
...
  diprec_api_internal_domain.UserResponse:
    properties:
      createdAt:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      patronymic:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  diprec_api_internal_domain.UserResponseWithCourses:
    properties:
      courses:
        items:
          $ref: '#/definitions/diprec_api_internal_domain.CourseResponse'
        type: array
      createdAt:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      patronymic:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  internal_transport_http_user.CreateUserDTO:
    properties:
      firstName:
        type: string
      lastName:
        type: string
      password:
        minLength: 8
        type: string
      patronymic:
        type: string
      username:
        type: string
    required:
    - firstName
    - lastName
    - password
    - username
    type: object
  internal_transport_http_user.LoginUserDTO:
    properties:
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_transport_http_user.RefreshUserDTO:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для дипломного проекта
  termsOfService: http://swagger.io/terms/
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для авторизации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_user.LoginUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      summary: Аутентификация пользователя
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh Token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_user.RefreshUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      summary: Обновить токены авторизации
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_user.CreateUserDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.AuthResponse'
        "400":
          description: Неверный формат запроса / тело запроса
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "401":
          description: Ошибка авторизации
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "409":
          description: Пользователь с таким именем уже существует
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      summary: Зарегистрировать нового пользователя
      tags:
      - Auth
...
  /user/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.UserResponseWithCourses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diprec_api_internal_domain.Error'
      security:
      - BearerAuth: []
      summary: Получение информации о текущем пользователе
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"